requirements.txt
pytorch
torchvision 
clip @ git+https://github.com/openai/CLIP.git